+++ src/Oro/Bundle/WsseAuthenticationBundle/Security/WsseTokenFactoryInterface.php
@@ -5,16 +5,16 @@ namespace Oro\Bundle\WsseAuthenticationBundle\Security;
 use Symfony\Component\Security\Core\Role\Role;
 
 /**
- * Interface for the factory which creates WsseToken.
+ * An interface for factories to create WsseToken.
  */
 interface WsseTokenFactoryInterface
 {
     /**
-     * @param string|object $user The username (like a nickname, email address, etc.),
-     *                                          or UserInterface instance or an object implementing a __toString method
-     * @param mixed $credentials This usually is the password of the user
-     * @param string $providerKey The provider key
-     * @param Role[]|string[] $roles An array of roles
+     * @param string|object $user        The username (like a nickname, email address, etc.),
+     *                                   or UserInterface instance or an object implementing a __toString method
+     * @param mixed         $credentials This usually is the password of the user
+     * @param string        $providerKey The provider key
+     * @param Role[]        $roles       An array of roles
      *
      * @return WsseToken
      */